+++ internal/filter/gateway/gateway_factory.go
@@ -7,7 +7,7 @@ import (
 
 type Factory interface {
 	Apply(configuration *config.FilterConfiguration) filter.Filter
-	GetOrder() int16
+	GetOrder() int
 }
 
 var Factories = make(map[string]Factory)
