+++ internal/filter/global/global.go
@@ -6,7 +6,7 @@ import (
 
 var Filters []filter.OrderedFilter
 
-func AddFilter(order int16, f filter.Filter) {
+func AddFilter(order int, f filter.Filter) {
 	orderedFilter := filter.NewOrderedFilter(order, f)
 	Filters = append(Filters, orderedFilter)
 }
